+++ mathics/core/atoms.py
@@ -225,6 +225,7 @@ Integer0 = Integer(0)
 Integer1 = Integer(1)
 Integer2 = Integer(2)
 Integer3 = Integer(3)
+Integer4 = Integer(4)
 Integer10 = Integer(10)